Fixed some inconsistencies with the package names
also removed some more redundant packages.
