+++ src/myn/patterns/Mediator/lab5/GreenCell.java
@@ -1,4 +1,4 @@
-package myn.patterns.Mediator.lab5;
+package myn.patterns.mediator.lab5;
 
 /**
  * Класс який наслідує абстрактний класс Cell
