+++ src/myn/patterns/Mediator/lab5/PlayField.java
@@ -1,10 +1,10 @@
+package myn.patterns.mediator.lab5;
+
 /**
  * @author Sienin Yuriy
  * @version 1.0
  * Інтерфейс PlayField, для медіатора
  */
-package myn.patterns.Mediator.lab5;
-
 interface PlayField {
     void changeActive(boolean changeActive, Cell cell);
 }
