+++ src/myn/patterns/Visitor/lab6_computer/IElementOfComputers.java
@@ -1,4 +1,4 @@
-package myn.patterns.Visitor.lab6_computer.src;
+package myn.patterns.visitor.lab6_computer;
 
 /**
  * @author Ivan Zaichenko
