+++ src/myn/patterns/Visitor/lab6_computer/Memory.java
@@ -1,5 +1,4 @@
-package myn.patterns.Visitor.lab6_computer.src;
-
+package myn.patterns.visitor.lab6_computer;
 /**
  * @author Ivan Zaichenko
  * @version 1.0
