+++ src/myn/patterns/Visitor/lab6_computer/Processor.java
@@ -1,18 +1,17 @@
-package myn.patterns.Visitor.lab6_computer.src;
-
+package myn.patterns.visitor.lab6_computer;
 /**
  * @author Ivan Zaichenko
  * @version 1.0
- * Клас VideoCard реалізує інтерфейс IElementOfComputers, створює елемет відеокарти
+ * Клас Processor реалізує інтерфейс IElementOfComputers, створює елемет процесора
  */
-public class VideoCard implements IElementOfComputers {
+public class Processor implements IElementOfComputers {
     /**
-     * Потужність відеокарти
+     * Потужність процесора
      */
     private double power;
 
     /**
-     * Повертає потужність відеокарти
+     * Повертає потужність процесора
      *
      * @return потужність всього компютера
      */
@@ -23,7 +22,7 @@ public class VideoCard implements IElementOfComputers {
     /**
      * Пустий конструктор
      */
-    public VideoCard() {
+    public Processor() {
         power = 100.5;
     }
 
@@ -32,7 +31,7 @@ public class VideoCard implements IElementOfComputers {
      *
      * @param power потужність елемента
      */
-    public VideoCard(double power) {
+    public Processor(double power) {
         this.power = power;
     }
