+++ src/myn/patterns/Visitor/visitor/Element.java
@@ -1,4 +1,4 @@
-package myn.patterns.Visitor.visitor;
+package myn.patterns.visitor.visitor;
 
 /**
  * @author Tsaruk Vova
