+++ src/myn/patterns/Visitor/visitor/ServerElement.java
@@ -1,4 +1,5 @@
-package myn.patterns.Visitor.visitor;
+package myn.patterns.visitor.visitor;
+
 
 /**
  * @author Tsaruk Vova
