+++ src/myn/patterns/adapter/dot_to_line/Dot.java
