+++ src/myn/patterns/adapter/dot_to_line/Line.java
