+++ src/myn/patterns/adapter/dot_to_line/LineToDot.java
