+++ src/myn/patterns/adapter/lab4/HundredWithLines.java
