+++ src/myn/patterns/adapter/lab4/IHundredWithLines.java
