+++ src/myn/patterns/adapter/lab4/ILine.java
