+++ src/myn/patterns/adapter/lab4/IRectangle.java
