+++ src/myn/patterns/adapter/lab4/Line.java
