+++ src/myn/patterns/adapter/lab4/Main.java
