+++ src/myn/patterns/adapter/lab4/Rectangle.java
