+++ src/myn/patterns/adapter/lab4/RectangleWithLines.java
