+++ src/myn/patterns/proxy/proxy/Color.java
@@ -1,4 +1,4 @@
-package myn.patterns.proxy.proxy.src.proxy;
+package myn.patterns.proxy.proxy;
 /**
  * @author Tsaruk Vova
  * Множина кольорів
