+++ src/myn/patterns/proxy/proxy/IManipulateImage.java
@@ -1,4 +1,4 @@
-package myn.patterns.proxy.proxy.src.proxy;
+package myn.patterns.proxy.proxy;
 /**
  * @author Tsaruk Vova
  * Інтерфейс для маніпулювання зображеннями
